2018.11.22

+++ economy/economy.h
@@ -118,7 +118,7 @@ namespace economy {
 	};
 
 	struct worked_instance {
-		int32_t worker_populations[max_worker_types] = { 0 };
+		float worker_populations[max_worker_types] = { 0 };
 		float production_scale = 1.0f;
 	};
